refactor: split AppModel into focused partial class files
Extract IAppSettings implementation to AppModel.AppSettings.cs,
INotificationSettings to AppModel.NotificationSettings.cs, and
IDeviceService to AppModel.DeviceService.cs. Trim main AppModel.cs
from 886 lines to 184 lines — keeping only singleton pattern,
constructor, InitializeMain, Dispose, and core events.

Fix missing using in INotificationSettings.cs for CachedSound type.

Phase 3 of splitting AppModel/IAppModel.
