+++ SoundSwitch/Model/AppModel.cs
@@ -84,418 +84,6 @@ public partial class AppModel : IAppModel
 
     public Services.AppSoundLockManager AppSoundLockManager { get; private set; }
 
-    public int BannerOnScreenTimeSecs
-    {
-        get => (int)BannerOnScreenTime.TotalSeconds;
-        set => BannerOnScreenTime = TimeSpan.FromSeconds(value);
-    }
-    public TimeSpan BannerOnScreenTime
-    {
-        get => AppConfigs.Configuration.BannerOnScreenTime;
-        set
-        {
-            if (value < TimeSpan.FromSeconds(2)) return;
-            if (value > TimeSpan.FromMinutes(1)) return;
-            var preValue = AppConfigs.Configuration.BannerOnScreenTime;
-            AppConfigs.Configuration.BannerOnScreenTime = value;
-            AppConfigs.Configuration.Save();
-            BannerSettingsChanged?.Invoke(this,
-                new BannerDataChangedEvent(BannerPosition, BannerPosition, preValue, value,
-                    BannerOpacityPercentage, BannerOpacityPercentage, BannerDisplayInfo, BannerDisplayInfo,
-                    MicrophoneMuteBanner, MicrophoneMuteBanner, MicrophoneUnmuteBanner, MicrophoneUnmuteBanner));
-        }
-    }
-
-    public int BannerOpacityPercentage
-    {
-        get => AppConfigs.Configuration.BannerOpacityPercentage;
-        set
-        {
-            if (value is < 10 or > 100) return;
-            var preValue = AppConfigs.Configuration.BannerOpacityPercentage;
-            AppConfigs.Configuration.BannerOpacityPercentage = value;
-            AppConfigs.Configuration.Save();
-            BannerSettingsChanged?.Invoke(this,
-                new BannerDataChangedEvent(BannerPosition, BannerPosition, BannerOnScreenTime, BannerOnScreenTime,
-                    preValue, value, BannerDisplayInfo, BannerDisplayInfo,
-                    MicrophoneMuteBanner, MicrophoneMuteBanner, MicrophoneUnmuteBanner, MicrophoneUnmuteBanner));
-        }
-    }
-
-    /// <summary>
-    /// How many notification to show at the same time
-    /// </summary>
-    public int MaxNumberNotification
-    {
-        get => AppConfigs.Configuration.MaxNumberNotification;
-        set
-        {
-            if (value is < 1 or > 100) return;
-            AppConfigs.Configuration.MaxNumberNotification = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    /// <summary>
-    /// Is there only 1 concurrent notification enabled ?
-    /// </summary>
-    public bool IsSingleNotification
-    {
-        get => AppConfigs.Configuration.MaxNumberNotification == 1;
-        set => MaxNumberNotification = value ? 1 : 5;
-    }
-
-    public CachedSound CustomNotificationSound
-    {
-        get
-        {
-            try
-            {
-                return _customNotificationCachedSound ??= new CachedSound(AppConfigs.Configuration.CustomNotificationFilePath);
-            }
-            catch (CachedSoundFileNotExistsException)
-            {
-                return null;
-            }
-        }
-        set
-        {
-            var oldSound = _customNotificationCachedSound;
-            _customNotificationCachedSound = value;
-            AppConfigs.Configuration.CustomNotificationFilePath = _customNotificationCachedSound?.FilePath;
-            AppConfigs.Configuration.Save();
-            CustomSoundChanged?.Invoke(this, new CustomSoundChangedEvent(oldSound, value));
-        }
-    }
-
-    public BannerDisplayInfo BannerDisplayInfo
-    {
-        get => AppConfigs.Configuration.BannerDisplayInfo;
-        set
-        {
-            var preValue = AppConfigs.Configuration.BannerDisplayInfo;
-            AppConfigs.Configuration.BannerDisplayInfo = value;
-            AppConfigs.Configuration.Save();
-            BannerSettingsChanged?.Invoke(this,
-                new BannerDataChangedEvent(BannerPosition, BannerPosition, BannerOnScreenTime, BannerOnScreenTime,
-                    BannerOpacityPercentage, BannerOpacityPercentage, preValue, value,
-                    MicrophoneMuteBanner, MicrophoneMuteBanner, MicrophoneUnmuteBanner, MicrophoneUnmuteBanner));
-        }
-    }
-
-    public bool NotificationAdvancedMode
-    {
-        get => AppConfigs.Configuration.NotificationAdvancedMode;
-        set
-        {
-            var previousSwitchDeviceNotification = AppConfigs.Configuration.SwitchDeviceNotification;
-            var previousSwitchProfileNotification = AppConfigs.Configuration.SwitchProfileNotification;
-            var previousMicrophoneMuteNotification = AppConfigs.Configuration.MicrophoneMuteNotification;
-
-            AppConfigs.Configuration.NotificationAdvancedMode = value;
-
-            if (!value)
-            {
-                AppConfigs.Configuration.SwitchProfileNotification = AppConfigs.Configuration.SwitchDeviceNotification;
-                AppConfigs.Configuration.MicrophoneMuteNotification = AppConfigs.Configuration.SwitchDeviceNotification;
-            }
-
-            AppConfigs.Configuration.Save();
-
-            NotificationSettingsChanged?.Invoke(this,
-                new NotificationSettingsUpdatedEvent(previousSwitchDeviceNotification, AppConfigs.Configuration.SwitchDeviceNotification,
-                    previousSwitchProfileNotification, AppConfigs.Configuration.SwitchProfileNotification,
-                    previousMicrophoneMuteNotification, AppConfigs.Configuration.MicrophoneMuteNotification));
-        }
-    }
-
-    public NotificationType SwitchDeviceNotification
-    {
-        get => AppConfigs.Configuration.SwitchDeviceNotification;
-        set
-        {
-            var previousSwitchDeviceNotification = AppConfigs.Configuration.SwitchDeviceNotification;
-            var previousSwitchProfileNotification = AppConfigs.Configuration.SwitchProfileNotification;
-            var previousMicrophoneMuteNotification = AppConfigs.Configuration.MicrophoneMuteNotification;
-
-            AppConfigs.Configuration.SwitchDeviceNotification = value;
-
-            if (!AppConfigs.Configuration.NotificationAdvancedMode)
-            {
-                AppConfigs.Configuration.SwitchProfileNotification = value;
-                AppConfigs.Configuration.MicrophoneMuteNotification = value;
-            }
-
-            AppConfigs.Configuration.Save();
-
-            NotificationSettingsChanged?.Invoke(this,
-                new NotificationSettingsUpdatedEvent(previousSwitchDeviceNotification, AppConfigs.Configuration.SwitchDeviceNotification,
-                    previousSwitchProfileNotification, AppConfigs.Configuration.SwitchProfileNotification,
-                    previousMicrophoneMuteNotification, AppConfigs.Configuration.MicrophoneMuteNotification));
-        }
-    }
-
-    public NotificationType SwitchProfileNotification
-    {
-        get => AppConfigs.Configuration.SwitchProfileNotification;
-        set
-        {
-            if (!AppConfigs.Configuration.NotificationAdvancedMode)
-            {
-                SwitchDeviceNotification = value;
-                return;
-            }
-
-            var preValue = AppConfigs.Configuration.SwitchProfileNotification;
-            AppConfigs.Configuration.SwitchProfileNotification = value;
-            AppConfigs.Configuration.Save();
-            NotificationSettingsChanged?.Invoke(this,
-                new NotificationSettingsUpdatedEvent(SwitchDeviceNotification, SwitchDeviceNotification,
-                    preValue, value, MicrophoneMuteNotification, MicrophoneMuteNotification));
-        }
-    }
-    public NotificationType MicrophoneMuteNotification
-    {
-        get => AppConfigs.Configuration.MicrophoneMuteNotification;
-        set
-        {
-            if (!AppConfigs.Configuration.NotificationAdvancedMode)
-            {
-                SwitchDeviceNotification = value;
-                return;
-            }
-
-            var preValue = AppConfigs.Configuration.MicrophoneMuteNotification;
-            AppConfigs.Configuration.MicrophoneMuteNotification = value;
-            AppConfigs.Configuration.Save();
-            NotificationSettingsChanged?.Invoke(this,
-                new NotificationSettingsUpdatedEvent(SwitchDeviceNotification, SwitchDeviceNotification,
-                    SwitchProfileNotification, SwitchProfileNotification, preValue, value));
-        }
-    }
-
-    public BannerPosition BannerPosition
-    {
-        get => AppConfigs.Configuration.BannerPosition;
-        set
-        {
-            var preValue = AppConfigs.Configuration.BannerPosition;
-            AppConfigs.Configuration.BannerPosition = value;
-            AppConfigs.Configuration.Save();
-            BannerSettingsChanged?.Invoke(this,
-                new BannerDataChangedEvent(preValue, value, BannerOnScreenTime, BannerOnScreenTime,
-                    BannerOpacityPercentage, BannerOpacityPercentage, BannerDisplayInfo, BannerDisplayInfo,
-                    MicrophoneMuteBanner, MicrophoneMuteBanner, MicrophoneUnmuteBanner, MicrophoneUnmuteBanner));
-        }
-    }
-
-    public Point CustomBannerPosition
-    {
-        get => AppConfigs.Configuration.CustomBannerPosition;
-        set
-        {
-            AppConfigs.Configuration.CustomBannerPosition = value;
-            AppConfigs.Configuration.Save();
-        }
-
-    }
-
-    /// <summary>
-    /// Current banner position implementation based on the BannerPosition setting
-    /// </summary>
-    public IPosition BannerPositionImpl => _bannerPositionFactory.Get(BannerPosition);
-
-    public MicrophoneMute MicrophoneMuteBanner
-    {
-        get => AppConfigs.Configuration.MicrophoneMuteBanner;
-        set
-        {
-            var prevValue = AppConfigs.Configuration.MicrophoneMuteBanner;
-            AppConfigs.Configuration.MicrophoneMuteBanner = value;
-            AppConfigs.Configuration.Save();
-            BannerSettingsChanged?.Invoke(this,
-                new BannerDataChangedEvent(BannerPosition, BannerPosition, BannerOnScreenTime, BannerOnScreenTime,
-                    BannerOpacityPercentage, BannerOpacityPercentage, BannerDisplayInfo, BannerDisplayInfo,
-                    prevValue, value, MicrophoneUnmuteBanner, MicrophoneUnmuteBanner));
-        }
-    }
-
-    public MicrophoneMute MicrophoneUnmuteBanner
-    {
-        get => AppConfigs.Configuration.MicrophoneUnmuteBanner;
-        set
-        {
-            var prevValue = AppConfigs.Configuration.MicrophoneUnmuteBanner;
-            AppConfigs.Configuration.MicrophoneUnmuteBanner = value;
-            AppConfigs.Configuration.Save();
-            BannerSettingsChanged?.Invoke(this,
-                new BannerDataChangedEvent(BannerPosition, BannerPosition, BannerOnScreenTime, BannerOnScreenTime,
-                    BannerOpacityPercentage, BannerOpacityPercentage, BannerDisplayInfo, BannerDisplayInfo,
-                    MicrophoneMuteBanner, MicrophoneMuteBanner, prevValue, value));
-        }
-    }
-
-    public IconDoubleClick IconDoubleClick
-    {
-        get => AppConfigs.Configuration.IconDoubleClick;
-        set
-        {
-            AppConfigs.Configuration.IconDoubleClick = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    /// <summary>
-    /// Beta or Stable channel.
-    /// </summary>
-    public bool IncludeBetaVersions
-    {
-        get => AppConfigs.Configuration.IncludeBetaVersions;
-        set
-        {
-            if (value != IncludeBetaVersions && _updateChecker != null)
-            {
-                _updateChecker.Beta = value;
-                CheckForUpdate();
-            }
-
-            AppConfigs.Configuration.IncludeBetaVersions = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public bool Telemetry
-    {
-        get => AppConfigs.Configuration.Telemetry;
-        set
-        {
-            AppConfigs.Configuration.Telemetry = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public bool QuickMenuEnabled
-    {
-        get => AppConfigs.Configuration.QuickMenuEnabled;
-        set
-        {
-            AppConfigs.Configuration.QuickMenuEnabled = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public bool KeepVolumeEnabled
-    {
-        get => AppConfigs.Configuration.KeepVolumeEnabled;
-        set
-        {
-            AppConfigs.Configuration.KeepVolumeEnabled = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public DeviceCollection<DeviceInfo> SelectedDevices
-    {
-        get
-        {
-            if (_selectedDevices != null)
-                return _selectedDevices;
-
-            return _selectedDevices = new DeviceCollection<DeviceInfo>(AppConfigs.Configuration.SelectedDevices.OrderBy(info => info.DiscoveredAt));
-        }
-    }
-
-    public IEnumerable<DeviceFullInfo> AvailablePlaybackDevices =>
-        AudioDeviceLister.GetDevices(DataFlow.Render, DeviceState.Active).IntersectWith(SelectedDevices);
-
-    public IEnumerable<DeviceFullInfo> AvailableRecordingDevices =>
-        AudioDeviceLister.GetDevices(DataFlow.Capture, DeviceState.Active).IntersectWith(SelectedDevices);
-
-    public bool SetCommunications
-    {
-        get => AppConfigs.Configuration.ChangeCommunications;
-        set
-        {
-            AppConfigs.Configuration.ChangeCommunications = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public UpdateMode UpdateMode
-    {
-        get => AppConfigs.Configuration.UpdateMode;
-        set
-        {
-            if (value != AppConfigs.Configuration.UpdateMode)
-            {
-                if (value != UpdateMode.Never)
-                    CheckForUpdate();
-
-                UpdateModeChanged?.Invoke(this, value);
-            }
-
-            AppConfigs.Configuration.UpdateMode = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public Language Language
-    {
-        get => AppConfigs.Configuration.Language;
-        set
-        {
-            AppConfigs.Configuration.Language = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public bool SwitchForegroundProgram
-    {
-        get => AppConfigs.Configuration.SwitchForegroundProgram;
-        set
-        {
-            AppConfigs.Configuration.SwitchForegroundProgram = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    public bool NotifyUsingPrimaryScreen
-    {
-        get => AppConfigs.Configuration.NotifyUsingPrimaryScreen;
-        set
-        {
-            AppConfigs.Configuration.NotifyUsingPrimaryScreen = value;
-            AppConfigs.Configuration.Save();
-        }
-    }
-
-    #region Misc settings
-
-    /// <summary>
-    ///     If the application runs at windows startup
-    /// </summary>
-    public bool RunAtStartup
-    {
-        get => AutoStart.IsAutoStarted();
-        set
-        {
-            Log.Information("Set AutoStart: {autostart}", value);
-            if (value)
-                AutoStart.EnableAutoStart();
-            else
-                AutoStart.DisableAutoStart();
-        }
-    }
-
-    public IAudioDeviceLister AudioDeviceLister { get; private set; }
-    public event EventHandler<NotificationSettingsUpdatedEvent> NotificationSettingsChanged;
-    public event EventHandler<BannerDataChangedEvent> BannerSettingsChanged;
-    public event EventHandler<CustomSoundChangedEvent> CustomSoundChanged;
-    public event EventHandler<UpdateMode> UpdateModeChanged;
-
-    #endregion
-
-
     /// <summary>
     ///     Initialize the Main class with Updater and Hotkeys
     /// </summary>
@@ -583,299 +171,9 @@ public partial class AppModel : IAppModel
         _initialized = true;
     }
 
-
-    /// <summary>
-    /// Handles microphone mute state changes and triggers appropriate notifications
-    /// </summary>
-    /// <param name="payload">The volume changed payload</param>
-    private void HandleMicrophoneMuteChanged(DeviceVolumeChangedPayload payload)
-    {
-        Log.Information("Microphone {DeviceName} mute state changed from {WasMuted} to {IsMuted}",
-            payload.Device.NameClean, payload.WasMuted, payload.IsMuted);
-
-        // Notify about the mute state change
-        _notificationManager.NotifyMuteChanged(payload.Device.Id, payload.Device.FriendlyName, payload.IsMuted);
-    }
-
-
-    private void InitUpdateChecker()
-    {
-#if DEBUG
-        const string url = "https://www.aaflalo.me/api.json";
-#else
-            const string url = "https://api.github.com/repos/Belphemur/SoundSwitch/releases";
-#endif
-        _updateChecker = new UpdateChecker(new Uri(url), AppConfigs.Configuration.IncludeBetaVersions);
-
-        _updateChecker.UpdateAvailable += (sender, @event) => NewVersionReleased?.Invoke(this,
-            new NewReleaseAvailableEvent(@event.AppRelease, AppConfigs.Configuration.UpdateMode));
-
-
-        JobScheduler.Instance.ScheduleJob(new CheckForUpdateRecurringJob(_updateChecker), CancellationToken.None, _updateScheduler);
-        Log.Information("Update checker initiated");
-    }
-
-    /// <summary>
-    /// For the app to check for update
-    /// </summary>
-    public void CheckForUpdate()
-    {
-        JobScheduler.Instance.ScheduleJob(new CheckForUpdateOnceJob(_updateChecker), CancellationToken.None, _updateScheduler);
-    }
-
-    public event EventHandler<DeviceListChanged> SelectedDeviceChanged;
+    public IAudioDeviceLister AudioDeviceLister { get; private set; }
     public event EventHandler<ExceptionEvent> ErrorTriggered;
     public event EventHandler<NewReleaseAvailableEvent> NewVersionReleased;
-
-    public event EventHandler<DeviceDefaultChangedEvent> DefaultDeviceChanged;
-
-    #region Selected devices
-
-    /// <summary>
-    /// Add the device to the Selected device list
-    /// </summary>
-    /// <param name="device"></param>
-    /// <returns></returns>
-    public bool SelectDevice(DeviceFullInfo device)
-    {
-        try
-        {
-            //Dont add device already selected
-            if (SelectedDevices.Contains(device))
-            {
-                return false;
-            }
-
-            device.DiscoveredAt = DateTime.UtcNow;
-            SelectedDevices.Add(device);
-            AppConfigs.Configuration.SelectedDevices = SelectedDevices.ToHashSet();
-        }
-        catch (ArgumentException)
-        {
-            return false;
-        }
-
-        SelectedDeviceChanged?.Invoke(this, new DeviceListChanged(SelectedDevices, device.Type));
-        AppConfigs.Configuration.Save();
-
-        return true;
-    }
-
-    /// <summary>
-    /// Add the device to the Selected device list
-    /// </summary>
-    /// <param name="device"></param>
-    /// <returns></returns>
-    public bool UnselectDevice(DeviceFullInfo device)
-    {
-        bool result;
-        try
-        {
-            result = SelectedDevices.Remove(device);
-            AppConfigs.Configuration.SelectedDevices = SelectedDevices.ToHashSet();
-        }
-        catch (ArgumentException)
-        {
-            return false;
-        }
-
-        if (result)
-        {
-            SelectedDeviceChanged?.Invoke(this,
-                new DeviceListChanged(SelectedDevices, device.Type));
-            AppConfigs.Configuration.Save();
-        }
-
-        return true;
-    }
-
-    #endregion
-
-    #region Hot keys
-
-    public bool SetHotkeyCombination(HotKey hotKey, HotKeyAction action, bool force = false)
-    {
-        var confHotKey = action switch
-        {
-            HotKeyAction.Playback => AppConfigs.Configuration.PlaybackHotKey,
-            HotKeyAction.Recording => AppConfigs.Configuration.RecordingHotKey,
-            HotKeyAction.Mute => AppConfigs.Configuration.MuteRecordingHotKey,
-            _ => throw new ArgumentOutOfRangeException(nameof(action), action, null)
-        };
-
-        if (!force && confHotKey == hotKey)
-        {
-            Log.Information("HotKey {action} already set {hotkeys}", action, confHotKey);
-            return true;
-        }
-
-        Log.Information("Unregister previous hotkeys {hotkeys}", confHotKey);
-        WindowsAPIAdapter.UnRegisterHotKey(confHotKey);
-        Log.Information("Unregistered previous hotkeys {hotkeys}", confHotKey);
-
-        if (!RegisterHotKey(hotKey)) return false;
-
-        Log.Information("New Hotkeys registered {hotkeys}", hotKey);
-
-        switch (action)
-        {
-            case HotKeyAction.Playback:
-                AppConfigs.Configuration.PlaybackHotKey = hotKey;
-                break;
-            case HotKeyAction.Recording:
-                AppConfigs.Configuration.RecordingHotKey = hotKey;
-                break;
-            case HotKeyAction.Mute:
-                AppConfigs.Configuration.MuteRecordingHotKey = hotKey;
-                break;
-            default:
-                throw new ArgumentOutOfRangeException(nameof(action), action, null);
-        }
-
-        AppConfigs.Configuration.Save();
-        return true;
-    }
-
-    private bool RegisterHotKey(HotKey hotkeys)
-    {
-        if (!hotkeys.Enabled || WindowsAPIAdapter.RegisterHotKey(hotkeys))
-            return true;
-
-        Log.Warning("Can't register new hotkeys {hotkeys}", hotkeys);
-        ErrorTriggered?.Invoke(this,
-            new ExceptionEvent(new Exception("Impossible to register HotKey: " + hotkeys)));
-        return false;
-    }
-
-
-    /// <summary>
-    /// Handles hotkey press events and performs the configured action for playback, recording, or mute hotkeys.
-    /// </summary>
-    /// <param name="sender">The event source.</param>
-    /// <param name="e">Key press event data whose <c>HotKey</c> is checked; unrelated hotkeys are ignored.</param>
-    private void HandleHotkeyPress(object sender, WindowsAPIAdapter.KeyPressedEventArgs e)
-    {
-        if (e.HotKey != AppConfigs.Configuration.PlaybackHotKey
-            && e.HotKey != AppConfigs.Configuration.RecordingHotKey
-            && e.HotKey != AppConfigs.Configuration.MuteRecordingHotKey)
-        {
-            Log.Debug("Not the registered Hotkeys {hotkeys}", e.HotKey);
-            return;
-        }
-
-        try
-        {
-            if (e.HotKey == AppConfigs.Configuration.PlaybackHotKey)
-                CycleActiveDevice(DataFlow.Render);
-            else if (e.HotKey == AppConfigs.Configuration.RecordingHotKey)
-                CycleActiveDevice(DataFlow.Capture);
-            else if (e.HotKey == AppConfigs.Configuration.MuteRecordingHotKey)
-                ToggleMicrophoneMute();
-        }
-        catch (Exception ex)
-        {
-            ErrorTriggered?.Invoke(this, new ExceptionEvent(ex));
-        }
-    }
-
-    #endregion
-
-    #region Microphone control
-
-    /// <summary>
-    /// Toggles the mute state of the default microphone
-    /// </summary>
-    /// <summary>
-    /// Toggles the system default microphone's mute state.
-    /// </summary>
-    /// <remarks>
-    /// If no default microphone is found or the operation fails, the method invokes the <c>ErrorTriggered</c> event and logs an error; on success it logs the new mute state.
-    /// </remarks>
-    public void ToggleMicrophoneMute()
-    {
-        var result = _microphoneMuteToggler.ToggleDefaultMute();
-        if (result == null)
-        {
-            ErrorTriggered?.Invoke(this, new ExceptionEvent(new Exception("No mic found or unable to toggle mute state")));
-            Log.Error("No mic found or unable to toggle mute state");
-        }
-        else
-        {
-            Log.Information("Microphone {DeviceName} mute state is now {IsMuted}", result.Value.Name, result.Value.MuteState);
-        }
-    }
-
-    /// <summary>
-    /// Toggles the mute state of the microphone
-    /// </summary>
-    /// <returns>Tuple with device name and mute state, null if no default microphone found or operation failed</returns>
-    public (string DeviceName, bool IsMuted)? SetMicrophoneMuteState(string deviceId, bool muteState)
-    {
-        var result = _microphoneMuteToggler.SetMicrophoneMuteState(deviceId, muteState);
-        if (result == null)
-        {
-            ErrorTriggered?.Invoke(this, new ExceptionEvent(new Exception("No mic found or unable to toggle mute state")));
-        }
-        return result;
-    }
-
-    /// <summary>
-    /// Sets the mute state of the default microphone
-    /// </summary>
-    /// <param name="muteState">The desired mute state</param>
-    /// <returns>Tuple with device name and mute state, null if no default microphone found or operation failed</returns>
-    public (string DeviceName, bool IsMuted)? SetMicrophoneMuteState(bool muteState)
-    {
-        var result = _microphoneMuteToggler.SetDefaultMuteState(muteState);
-        if (result == null)
-        {
-            ErrorTriggered?.Invoke(this, new ExceptionEvent(new Exception("No mic found or unable to set mute state")));
-        }
-        return result;
-    }
-
-    #endregion
-
-    #region Active device
-
-    /// <summary>
-    ///     Attempts to set active device to the specified name
-    /// </summary>
-    /// <param name="device"></param>
-    public bool SetActiveDevice(DeviceInfo device)
-    {
-        try
-        {
-            return _deviceCyclerManager.SetAsDefault(device);
-        }
-        catch (Exception ex)
-        {
-            ErrorTriggered?.Invoke(this, new ExceptionEvent(ex));
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    ///     Cycles the active device to the next device. Returns true if succesfully switched (at least
-    ///     as far as we can tell), returns false if could not successfully switch. Throws NoDevicesException
-    ///     if there are no devices configured.
-    /// </summary>
-    public bool CycleActiveDevice(DataFlow type)
-    {
-        try
-        {
-            return _deviceCyclerManager.CycleDevice(type);
-        }
-        catch (Exception exception)
-        {
-            ErrorTriggered?.Invoke(this, new ExceptionEvent(exception));
-        }
-
-        return false;
-    }
-
-    #endregion
 
     public void Dispose()
     {
